finish gossipProtocol

+++ src/sims/collect/BroadcastObserver.java
@@ -1,7 +1,6 @@
 package sims.collect;
 
 import java.util.HashMap;
-import java.util.Map;
 
 import peersim.config.Configuration;
 import peersim.core.Control;
@@ -51,7 +50,7 @@ public boolean execute() {
 static public void handleSendMsg(int protocolID, Node from, Node to, Message msg) {
    msgStats.add(msg.GetSize());
    // nodeStats.add(to.getIndex());
-   Integer recv = (Integer)nodeRecvs.getOrDefault(to.getID(), 0);
+   Integer recv = (Integer)nodeRecvs.getOrDefault((int)to.getID(), 0);
    nodeRecvs.put( (int)to.getID(), recv+1);
 }
